+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This applies from version 0.5.0 onwards, as some versions before that have broke
 ### Added
 
 - Add "Stop editing" button.
+- Add a "Copy to clipboard" button for each item.
 
 ### Changed